fix(SPM): Move package definition so it's resolvable by Xcode, and adapt the Example project
